Day 17/server/requirements.txt
fastapi
uvicorn
python-dotenv
assemblyai
websockets
